CLIENT-2112 Reset the query “done” indicator before checking for errors in PartitionTracker.IsComplete().
When a retryable error (like INDEX_NOTFOUND) occurs on all nodes, the query iteration is marked complete. Later, maxRetries is exceeded, but the query terminate does not happen because the query was already marked complete. The fix is to reset the query “done” indicator before checking for errors in PartitionTracker.IsComplete().